+++ src/structs/cmd_utils.c
@@ -6,7 +6,7 @@
 /*   By: mortins- <mortins-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/08 17:08:00 by ddiniz-m          #+#    #+#             */
-/*   Updated: 2023/10/11 17:38:31 by mortins-         ###   ########.fr       */
+/*   Updated: 2023/10/12 16:50:57 by mortins-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -81,26 +81,4 @@ int	cmd_count(char **arr)
 			i++;
 	}
 	return (cmd_n);
-}
-
-//Returns lenght of string when it reaches c
-int	strlen_chr(char *str, char c)
-{
-	int	i;
-
-	i = 0;
-	while (str[i] && str[i] != c)
-		i++;
-	return (i);
-}
-
-//Compares s1 and s2 until s2 reaches c
-int	strcmp_chr(char *s1, char *s2, char c)
-{
-	int	i;
-
-	i = 0;
-	while (i < strlen_chr(s2, c) - 1)
-		i++;
-	return (s1[i] - s2[i]);
 }
